AWS SSO doesn't support resource-based policies
At least according to its documentation [here](https://docs.aws.amazon.com/singlesignon/latest/userguide/iam-auth-access-overview.html#accesscontrolresourcebased):


AWS Documentation » AWS Single Sign-On » User Guide » Authentication and Access Control for AWS SSO » Overview of Managing Access Permissions to Your AWS SSO Resources
Overview of Managing Access Permissions to Your AWS SSO Resources
Every AWS resource is owned by an AWS account, and permissions to create or access the resources are governed by permissions policies. An account administrator can attach permissions policies to IAM identities (that is, users, groups, and roles). Some services (such as AWS Lambda) also support attaching permissions policies to resources.

Note

An account administrator (or administrator user) is a user with administrator privileges. For more information, see IAM Best Practices in the IAM User Guide.

When granting permissions, you decide who is getting the permissions, the resources they get permissions for, and the specific actions that you want to allow on those resources.

Topics

AWS SSO Resources and Operations
Understanding Resource Ownership
Managing Access to Resources
Specifying Policy Elements: Actions, Effects, Resources, and Principals
Specifying Conditions in a Policy
AWS SSO Resources and Operations
In AWS SSO, the primary resources are application instances, profiles, and permission sets.

Understanding Resource Ownership
A resource owner is the AWS account that created a resource. That is, the resource owner is the AWS account of the principal entity (the account, an IAM user, or an IAM role) that authenticates the request that creates the resource. The following examples illustrate how this works:

If the AWS account root user creates an AWS SSO resource, such as an application instance or permission set, your AWS account is the owner of that resource.

If you create an IAM user in your AWS account and grant that user permissions to create AWS SSO resources, the user can then create AWS SSO resources. However, your AWS account, to which the user belongs, owns the resources.

If you create an IAM role in your AWS account with permissions to create AWS SSO resources, anyone who can assume the role can create AWS SSO resources. Your AWS account, to which the role belongs, owns the AWS SSO resources.

Managing Access to Resources
A permissions policy describes who has access to what. The following section explains the available options for creating permissions policies.

Note

This section discusses using IAM in the context of AWS SSO. It doesn't provide detailed information about the IAM service. For complete IAM documentation, see What Is IAM? in the IAM User Guide. For information about IAM policy syntax and descriptions, see AWS IAM Policy Reference in the IAM User Guide.

Policies that are attached to an IAM identity are referred to as identity-based policies (IAM policies). Policies that are attached to a resource are referred to as resource-based policies. AWS SSO supports only identity-based policies (IAM policies).

Topics

Identity-Based Policies (IAM Policies)
Resource-Based Policies
Identity-Based Policies (IAM Policies)
You can attach policies to IAM identities. For example, you can do the following:

Attach a permissions policy to a user or a group in your account – An account administrator can use a permissions policy that is associated with a particular user to grant permissions for that user to add an AWS SSO resource, such as a new application.

Attach a permissions policy to a role (grant cross-account permissions) – You can attach an identity-based permissions policy to an IAM role to grant cross-account permissions. For example, the administrator in Account A can create a role to grant cross-account permissions to another AWS account (for example, Account B) or an AWS service as follows:

Account A administrator creates an IAM role and attaches a permissions policy to the role that grants permissions to resources in Account A.

Account A administrator attaches a trust policy to the role identifying Account B as the principal who can assume the role.

Account B administrator can then delegate permissions to assume the role to any users in Account B. Doing this allows users in Account B to create or access resources in Account A. The principal in the trust policy can also be an AWS service principal if you want to grant an AWS service permissions to assume the role.

For more information about using IAM to delegate permissions, see Access Management in the IAM User Guide.

The following permissions policy grants permissions to a user to run all of the actions that begin with List. These actions show information about an AWS SSO resource, such as an application instance or permissions set. Note that the wildcard character (*) in the Resource element indicates that the actions are allowed for all AWS SSO resources that are owned by the account.

{
   "Version":"2012-10-17",
   "Statement":[
      {
         "Effect":"Allow",
         "Action":"sso:List*",
         "Resource":"*"
      }
   ]
}
For more information about using identity-based policies with AWS SSO, see Using Identity-Based Policies (IAM Policies) for AWS SSO. For more information about users, groups, roles, and permissions, see Identities (Users, Groups, and Roles) in the IAM User Guide.

> Other services, such as Amazon S3, also support resource-based permissions policies. For example, you can attach a policy to an S3 bucket to manage access permissions to that bucket. AWS SSO doesn't support resource-based policies.

This change fixes the table to match the service's own documentation.